trabajo_final/requirements.txt
django==2.2.5
ipython==7.7.0
jupyter==1.0.0
requests==2.22.0
djangorestframework
markdown
django-filter
